chore: move ConnectorStatus to common and test it
