+++ toucan_connectors/mongo/mongo_connector.py
@@ -9,10 +9,9 @@ from bson.son import SON
 from cached_property import cached_property
 from pydantic import Field, SecretStr, create_model, validator
 
-from toucan_connectors.common import nosql_apply_parameters_to_query
+from toucan_connectors.common import ConnectorStatus, nosql_apply_parameters_to_query
 from toucan_connectors.mongo.mongo_translator import MongoConditionTranslator
 from toucan_connectors.toucan_connector import (
-    ConnectorStatus,
     DataSlice,
     ToucanConnector,
     ToucanDataSource,
